Switch to casting that is easier to read
This is primarily to make the code easier to read than using "static_cast<type>(exp)".

In Debug builds, to_int and to_char will check to be certain that a larger size can be reduced without loss of data.

+++ src/ui/embedimg.cpp
@@ -496,7 +496,7 @@ void EmbedImage::ImgageInHeaderOut()
             auto& line = file.addEmptyLine();
             for (; pos < buf_size && line.size() < 116; ++pos)
             {
-                line << static_cast<int>(buf[pos]) << ',';
+                line << buf[pos] << ',';
             }
         }
     }
